+++ frontend/src/app/layout.tsx
@@ -1,0 +1,31 @@
+import type { Metadata } from "next";
+import { Poppins } from "next/font/google";
+import "./globals.css";
+
+const poppins = Poppins({
+	weight: ["400", "500", "600", "700"],
+	subsets: ["latin", "latin-ext"],
+	variable: "--font-poppins",
+	display: "swap",
+	preload: true,
+});
+
+export const metadata: Metadata = {
+	title: "NegoProx - Conectando negocios",
+	description:
+		"Conectando negocios, desde la comodidad de tu hogar y facilitando la comunicación con tus clientes.",
+};
+
+export default function RootLayout({
+	children,
+}: Readonly<{
+	children: React.ReactNode;
+}>) {
+	return (
+		<html lang="es">
+			<body className={`${poppins.variable} antialiased`}>
+				<div>{children}</div>
+			</body>
+		</html>
+	);
+}
